+++ RELEASE_NOTES.md
@@ -3,6 +3,8 @@
 * Fix `Message` colors properties
 * Add `Message` documentation
 * Add `Size` properties to `Message`
+* `Nav` has been dreprecated in Bulma. Users should use `Navbar` from now.
+* Remove `FSharp.NET.Sdk` from the dependencies.
 
 ### 0.1.0-alpha9